requirements.txt
pythondialog==3.5.3
